Removed Old Dash 1.1 Sonic Content
Removed and reorganized the Characters folder to remove the old 1.1 Sonic content. Old gens animations and skeletal mesh are removed alongside the old materials and textures.
Particles and their meshes materials and textures in Characters>Sonic are put in a folder with OLD tagged to the end of it in the NovaSonic folder.

Some stuff outside of the Characters folder are edited.
Empire City redirectors are edited on accident.
Some defaults in the footstep struct, brake struct, etc are edited.