Laser Fire Input
Laser now fires off if the Special Action input is NOT pressed. This is done via a timer event that checks if a bool in the player pawn (which is set to true/false if special key is pressed/released) is not true. If the bool is not true, it performs the fire event and ends the timer.