New Main Menu + Wisp HUD Anim Hotfix
Redid the whole main menu from scratch, salvaging any concepts/systems from the old one if they were any good.
The only thing left is the background for it and then the options menu.
Also fixed a small oversight with the HUD animation for using a wisp not properly stopping itself when the wisp ends.