Main Menu State-Based Transitions
Added state-based transitions to the main menu. What this means is that when the main menu is opened, the manager actor checks what the current menu state should be (from game instance) and then does something based off of that. This means that now the player can go right to the level select after finishing a level instead of being shoved back all the way to the title screen.